Final del tutorial parte de backend, con ajustes y modificaciones.

+++ tutorial/src/main/java/com/ccsw/tutorial/client/ClientServiceImpl.java
@@ -3,6 +3,8 @@ package com.ccsw.tutorial.client;
 
 import com.ccsw.tutorial.client.model.Client;
 import com.ccsw.tutorial.client.model.ClientDto;
+import com.ccsw.tutorial.loan.LoanRepository;
+import com.ccsw.tutorial.loan.model.Loan;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,6 +16,9 @@ public class ClientServiceImpl implements ClientService {
 
     @Autowired
     ClientRepository clientRepository;
+
+    @Autowired
+    private LoanRepository loanRepository;
 
     @Override
     public Client get(Long id) {
@@ -51,10 +56,15 @@ public class ClientServiceImpl implements ClientService {
     @Override
     public void delete(Long id) throws Exception {
 
-
         if (this.get(id) == null) {
             throw new Exception("Not exists");
         }
+
+        List<Loan> clientLoans = this.loanRepository.findByClientId(id);
+        if (!clientLoans.isEmpty()) {
+            throw new RuntimeException("No se puede eliminar el cliente porque tiene préstamos asociados en el sistema.");
+        }
+
         this.clientRepository.deleteById(id);
     }
 }
